+++ pipelines/control__profiling_baseline/CHANGELOG.md
@@ -1,0 +1,9 @@
+# Changelog - control__profiling_baseline
+
+## [0.1.0] - 2026-05-19
+
+### Adicionado
+
+- Flow `control__profiling_baseline`: flow vazio instrumentado com
+  `profile_resources` para medir o consumo de recursos de um Prefect flow
+  sem operações. Serve como linha-base.
