+++ pipelines/control__profiling_baseline/Dockerfile
@@ -4,9 +4,11 @@ LABEL org.opencontainers.image.source=https://github.com/rj-smtr/pipelines_v3
 
 WORKDIR /opt/prefect/pipelines_v3
 
+# Sem COPY ./queries: este flow é apenas baseline e não usa dbt.
+
 COPY ./pyproject.toml ./uv.lock /opt/prefect/pipelines_v3/
 
-COPY ./pipelines/control__profiling_dbt ./pipelines/control__profiling_dbt/
+COPY ./pipelines/control__profiling_baseline ./pipelines/control__profiling_baseline/
 
 COPY ./pipelines/common/ ./pipelines/common/
