+++ pipelines/control__profiling_baseline/prefect.yaml
@@ -1,0 +1,37 @@
+name: control__profiling_baseline
+prefect-version: 3.4.9
+
+build:
+  - prefect.deployments.steps.run_shell_script:
+      id: get-commit-hash
+      script: git rev-parse --short HEAD
+      stream_output: false
+  - prefect_docker.deployments.steps.build_docker_image:
+      id: build-image
+      requires: prefect-docker>=0.6.5
+      image_name: ghcr.io/rj-smtr/pipelines_v3/deployments
+      tag: "control__profiling_baseline-{{ get-commit-hash.stdout }}"
+      dockerfile: pipelines/control__profiling_baseline/Dockerfile
+
+push:
+  - prefect_docker.deployments.steps.push_docker_image:
+      requires: prefect-docker>=0.6.5
+      image_name: "{{ build-image.image_name }}"
+      tag: "{{ build-image.tag }}"
+
+pull:
+  - prefect.deployments.steps.set_working_directory:
+      directory: /opt/prefect/pipelines_v3
+
+deployments:
+  - name: rj-control--profiling-baseline--staging
+    version: "{{ build-image.tag }}"
+    entrypoint: pipelines/control__profiling_baseline/flow.py:control__profiling_baseline
+    work_pool:
+      name: smtr-pool
+      work_queue_name: default
+      job_variables:
+        image: "{{ build-image.image_name }}:{{ build-image.tag }}"
+        command: uv run --package control__profiling_baseline -- prefect flow-run execute
+        secretName: prefect-jobs-secrets-staging
+        image_pull_policy: Always
