+++ pipelines/control__profiling_dbt_runtime/constants.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-"""Constantes do pipeline de profiling/teste de download de queries"""
+"""Constantes do pipeline de teste de download de queries em runtime"""
 
 # Repositório de onde a pasta `queries` é baixada em runtime
 GIT_REPO_URL = "https://github.com/RJ-SMTR/pipelines_v3.git"
@@ -9,6 +9,3 @@ DEFAULT_GIT_REF = "master"
 
 # Nome da variável de ambiente que contém o token do GitHub (opcional, para repo privado)
 GITHUB_TOKEN_ENV = "GITHUB_TOKEN"
-
-# Intervalo (segundos) entre amostras do profiler de recursos
-PROFILE_SAMPLE_INTERVAL = 0.5
